+++ app/src/components/NewGame/NewGame.jsx
@@ -9,7 +9,7 @@ import * as Yup from 'yup'
 import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useGameStore } from '../../store/gameStore/useGameStore'
-import api from '../../tools/Api/Api'
+import api from '../../services/Api/Api'
 import { MemoButton } from '../../ui/MemoButton/MemoButton'
 import { Loader } from '../Loader/Loader'
 import { Modal } from '../Modal/Modal'
