+++ app/src/components/SignIn/SignIn.jsx
@@ -7,7 +7,7 @@ import { MemoButton } from '../../ui/MemoButton/MemoButton'
 import styles from './signin.module.scss'
 import { validationSchemaSignIn } from './validationSchema/validationSchema'
 import { notifyError, notifySuccess } from '../../tools/toaster/toaster.js'
-import api from '../../tools/Api/Api.js'
+import api from '../../services/Api/Api.js'
 import fullLogo from '../../resources/images/logo/logo_full.png'
 
 export const SignIn = () => {
